Add files via upload
Here is the final documentation of my thesis for my graduate school as well as the detailed description to the analysis performed. For the code used, if you have access to the free application of the latest Houdini build you can open up the .hip file or you can go to the scripts branch under the same repository to have a look at the code used.